WFLY-16322 Ensure proper handling of content type availability in VirtualFileURLConnection
In the JDK 17.0.3 release (and other releases, i.e. 11.0.16, 13.0.12,
15.0.8)), the backport of "JDK-8273655 content-types.properties files
are missing some common types" breaks VFS in VirtualFileUrlConnnection.

Instead of returning the VirtualFile as expected (i.e. previous behavior
prior to JDK-8274655) it now returns the input stream from the super
class, causing issues in client libraries that depend on getting a
VirtualFile.

This commit retains the old behavior, and introduces a check for the
content type "application/java-archive" as specified in JDK-8273655's
properties files.